fix silent output bug #2 for low latency notebook for Python < 3

+++ gccNMF/wavfile.py
@@ -34,7 +34,7 @@ def wavread(filePath):
     return samples_float32.T, sampleRate
     
 def wavwrite(samples_float32, filePath, sampleRate):
-    samples_pcm = float2pcm(samples_float32)
+    samples_pcm = float2pcm( samples_float32.astype(np.float32) )
     wavfile.write( filePath, sampleRate, samples_pcm.T )
     
 """
